Fix border guards getting stuck
