+++ game/src/main/kotlin/content/area/misthalin/BorderGuard.kts
@@ -9,22 +9,18 @@ import world.gregs.voidps.engine.entity.obj.ObjectLayer
 import world.gregs.voidps.engine.entity.worldSpawn
 import world.gregs.voidps.engine.inject
 import world.gregs.voidps.type.Distance.nearestTo
-import world.gregs.voidps.type.Tile
-import world.gregs.voidps.type.Zone
 import world.gregs.voidps.type.area.Rectangle
 import kotlin.collections.set
 
 val objects: GameObjects by inject()
 val areas: AreaDefinitions by inject()
 
-val borders = mutableMapOf<Zone, Rectangle>()
 val guards = mutableMapOf<Rectangle, List<GameObject>>()
 
 worldSpawn {
     for (border in areas.getTagged("border")) {
         val passage = border.area as Rectangle
         for (zone in passage.toZones()) {
-            borders[zone] = passage
             guards[passage] = zone.toRectangle().mapNotNull {
                 val obj = objects.getLayer(it, ObjectLayer.GROUND)
                 if (obj != null && obj.id.startsWith("border_guard")) obj else null
@@ -35,9 +31,9 @@ worldSpawn {
 
 enterArea("border_guard*") {
     val border = area as Rectangle
-    if (player.steps.destination in border) {
+    if (player.steps.destination in border || player.steps.isEmpty()) {
         val tile = border.nearestTo(player.tile)
-        val endSide = getOppositeSide(border, tile)
+        val endSide = Border.getOppositeSide(border, tile)
         player.walkTo(endSide, noCollision = true, forceWalk = true)
     } else {
         player.steps.update(noCollision = true, noRun = true)
@@ -62,11 +58,4 @@ fun changeGuardState(guards: List<GameObject>, raise: Boolean) {
             raised[guard] = raise
         }
     }
-}
-
-// Longest axis determines direction, current location above is underside else above
-fun getOppositeSide(border: Rectangle, tile: Tile) = if (border.height > border.width) {
-    tile.copy(y = if (tile.y > border.minY) border.minY - 1 else border.maxY + 1)
-} else {
-    tile.copy(x = if (tile.x > border.minX) border.minX - 1 else border.maxX + 1)
 }
